formatting and rework to be drivers only
