+++ perception/webcam_sensor.py
@@ -1,11 +1,11 @@
 import cv2
 import logging
-import numpy as np
 import os
 import subprocess
-import shlex
 
-from . import CameraSensor, ColorImage, CameraIntrinsics
+from autolab_core import ColorImage, CameraIntrinsics
+
+from .camera_sensor import CameraSensor
 
 
 class WebcamSensor(CameraSensor):
@@ -49,7 +49,7 @@ class WebcamSensor(CameraSensor):
 
     @property
     def color_intrinsics(self):
-        """CameraIntrinsics : The camera intrinsics for the PhoXi Greyscale camera."""
+        """Camera intrinsics for the PhoXi Greyscale camera."""
         return self._camera_intr
 
     @property
@@ -101,37 +101,43 @@ class WebcamSensor(CameraSensor):
         return True
 
     def frames(self, most_recent=False):
-        """Retrieve a new frame from the PhoXi and convert it to a ColorImage,
-        a DepthImage, and an IrImage.
+        """Retrieve a new frame from the Webcam and convert it to a
+        ColorImage and DepthImage pair.
 
         Parameters
         ----------
         most_recent: bool
-            If true, the OpenCV buffer is emptied for the webcam before reading the most recent frame.
+            If true, the OpenCV buffer is emptied for the webcam
+            before reading the most recent frame.
 
         Returns
         -------
-        :obj:`tuple` of :obj:`ColorImage`, :obj:`DepthImage`, :obj:`IrImage`, :obj:`numpy.ndarray`
-            The ColorImage, DepthImage, and IrImage of the current frame.
+        :obj:`tuple` of :obj:`ColorImage`, :obj:`DepthImage`
+            The ColorImage and DepthImage of the current frame.
         """
         if most_recent:
-            for i in range(4):
+            for _ in range(4):
                 self._cap.grab()
-        for i in range(1):
+        for _ in range(1):
             if self._adjust_exposure:
                 try:
-                    command = "v4l2-ctl -d /dev/video{} -c exposure_auto=1 -c exposure_auto_priority=0 -c exposure_absolute=100 -c saturation=60 -c gain=140".format(
-                        self._device_id
-                    )
+                    command = [
+                        "v4l2-ctl",
+                        "-d /dev/video{}".format(self._device_id),
+                        "-c exposure_auto=1",
+                        "-c exposure_auto_priority=0",
+                        "-c exposure_absolute=100",
+                        "-c saturation=60" "-c gain=140",
+                    ]
                     FNULL = open(os.devnull, "w")
                     subprocess.call(
-                        shlex.split(command),
+                        command,
                         stdout=FNULL,
                         stderr=subprocess.STDOUT,
                     )
-                except:
+                except subprocess.SubprocessError:
                     pass
-            ret, frame = self._cap.read()
+            _, frame = self._cap.read()
         rgb_data = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
 
-        return ColorImage(rgb_data, frame=self._frame), None, None
+        return ColorImage(rgb_data, frame=self._frame), None
